+++ tests/lint/contents/invalid/Contents.m
@@ -1,0 +1,12 @@
+% Package for dealing with potatos
+%
+% Functions to represent potatos
+%   newpot  - Create a new potato
+%   copypot - Copy a potato
+%
+% Functions to explain potatos
+%   wtp     - Why the potato???
+
+disp("Potato");
+
+% (c) Copyright 2021 Florian Schanda