+++ CHANGELOG.md
@@ -54,6 +54,9 @@ Not quite compatible with Octave yet. See #43 [octave support](https://github.co
   check uses a much broader list of built-in functions, than the check
   that looks at assignments.
 
+* New check for MH Lint to make sure that any file called `Contents.m`
+  only contains comments.
+
 ### 0.9.17
 
 * Updated language support to MATLAB 2021a. The new feature supported